+++ packages/video/src/index.ts
@@ -1,6 +1,6 @@
 "use client";
 
-export { VideoProvider } from "./wrapper";
+export { VideoProvider, useVideo } from "./wrapper";
 export { Video } from "./video";
 export * from "./components";
 export * from "./types";